move the block headers inside blocklib to make blocklib portable
