+++ blocklib/CMakeLists.txt
@@ -1,1 +1,2 @@
+add_subdirectory(common)
 add_subdirectory(blocks)
