+++ blocklib/blocks/include/gnuradio/blocklib/blocks/vector_source.hpp
@@ -1,7 +1,7 @@
 #ifndef INCLUDED_VECTOR_SOURCE_HPP
 #define INCLUDED_VECTOR_SOURCE_HPP
 
-#include <gnuradio/sync_block.hpp>
+#include <gnuradio/blocklib/sync_block.hpp>
 
 namespace gr {
 namespace blocks {
