+++ blocklib/blocks/lib/CMakeLists.txt
@@ -11,11 +11,10 @@ target_include_directories(gnuradio-blocklib-blocks
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../include>
   PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/../include/gnuradio/blocklib/blocks
-  ${PROJECT_SOURCE_DIR}/runtime/include 
+  ${PROJECT_SOURCE_DIR}/blocklib/common/include 
   )
 
   # TODO - learn how to include against runtime with INSTALL_INTERFACE
 
   target_link_libraries(gnuradio-blocklib-blocks PUBLIC
-    Volk::volk
-    gnuradio-runtime)
+    Volk::volk)
