+++ blocklib/common/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_subdirectory(include/gnuradio/blocklib)
+add_subdirectory(lib)
