+++ blocklib/common/include/gnuradio/blocklib/CMakeLists.txt
@@ -2,11 +2,13 @@
 # Install header files
 ########################################################################
 install(FILES
-  api.h
-  attributes.h
-  flowgraph.hpp
-  scheduler.hpp
-  DESTINATION ${GR_INCLUDE_DIR}/gnuradio
+  block.hpp
+  block_callbacks.hpp
+  io_signature.hpp
+  sync_block.hpp
+  tag.hpp
+  types.hpp
+  DESTINATION ${GR_INCLUDE_DIR}/gnuradio/blocklib/
 )
 
 message(STATUS ${CMAKE_INSTALL_INCLUDEDIR})
